pyproject.toml
[project]
name = "microremote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "keyboard>=0.13.5",
    "pyserial>=3.5",
]
